osa4

+++ osa4/blogilista/controllers/blogs.js
@@ -1,14 +1,23 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const jwt = require('jsonwebtoken')
+const { userExtractor } = require('../utils/middleware')
 
-blogsRouter.get('/', async (request, response) => { 
-  const blogs = await Blog.find({})
+blogsRouter.get('/', async (request, response) => {
+  const blogs = await Blog
+    .find({}).populate('user', { username: 1, name: 1 })
+
   response.json(blogs)
 })
 
-
-blogsRouter.post('/', async (request, response, next) => {
+blogsRouter.post('/', userExtractor, async (request, response, next) => {
   const body = request.body
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {    
+    return response.status(401).json({ error: 'token invalid' })  
+  }
+  const user = request.user
+
   if (body.title === undefined) {
     return response.status(400).json({ error: "Missing title" })
   } else if (body.url === undefined) {
@@ -18,18 +27,27 @@ blogsRouter.post('/', async (request, response, next) => {
     title: body.title,
     author: body.author,
     url: body.url,
-    likes: body.likes || 0
+    likes: body.likes || 0,
+    user: user._id
   })
 
     const savedblog = await blog.save({})
+    user.blogs = user.blogs.concat(savedblog._id)  
+    await user.save()
     response.status(201).json(savedblog)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+blogsRouter.delete('/:id', userExtractor, async (request, response) => {
+  const user = request.user
+  const blog = await Blog.findById(request.params.id)
+  
+  if (user.id.toString() === blog.user.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    return response.status(401).json({ error: 'user not correct' })  
+  }
 })
-
 
 blogsRouter.put('/:id', async (request, response) => {
   const body = request.body
